+++ src/main/java/com/example/be_swp/DTOs/PaymentDTO.java
@@ -18,7 +18,7 @@ public class PaymentDTO {
 
     private int paymentId;
 
-    private int appointmentId;
+    private Long appointmentId;
 
     private int paymentMethod;
